Ai-Answer-Sheet-Validation-main/backend/requirements.txt
PyMuPDF
aiohttp
nltk
sentence_transformers
pymongo
google-generativeai
pyppeteer 
